Inclusao Dockerfile e scripts de build

+++ src/main/java/com/euax/desafio/resources/ProjectResource.java
@@ -14,12 +14,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import com.euax.desafio.domain.Project;
-import com.euax.desafio.domain.Task;
 import com.euax.desafio.dto.ProjectDTO;
-import com.euax.desafio.dto.TaskDTO;
 import com.euax.desafio.services.ProjectService;
 
 
@@ -89,6 +88,11 @@ public class ProjectResource {
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Void> delete (@PathVariable Integer id){
+		
+		Project project = service.findWithoutValidation(id);
+		
+		if (project == null)
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
 		
 		service.delete(id);
